悬浮画板导入图片输入数值加入数学表达式解析计算 版本号更换成日期int 格式化工程代码

+++ app/src/main/java/pers/zhc/tools/functiondrawing/FunctionDrawing.java
@@ -48,7 +48,7 @@ public class FunctionDrawing extends BaseActivity {
         fs = new FourierSeries(intent.getIntExtra("FS_T", 30)) {
             @Override
             public double f_f(double x) {
-                return ((double) FunctionDrawingBoard.functionInterface.f(((float) x)));
+                return FunctionDrawingBoard.functionInterface.f(((float) x));
             }
         };
         fs.definite.n = 20000;
